backend/requirements.txt
azure-ai-formrecognizer
Flask
Flask-Cors
gunicorn
Jinja2
pillow
python-dotenv
cloudinary